refactor: move color definitions into design
