+++ src/view/design/AppHeader.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { Appbar, AppbarHeaderProps, Theme, withTheme } from 'react-native-paper'
 import { ViewStyle } from 'react-native'
-import { headerColor } from '../color'
+import { headerColor } from './color'
 
 const AppHeaderImpl: React.FC<AppbarHeaderProps & { theme: Theme }> = ({
   children,
